+++ Runtime/Scripts/Data/KnotEndpoint.cs
@@ -13,12 +13,20 @@ namespace Knot.OpenAI
         }
         [SerializeField] private string _uri;
 
+        public int Timeout
+        {
+            get => _timeout;
+            set => _timeout = value;
+        }
+        [SerializeField] private int _timeout = 10;
+
 
         public KnotEndpoint() { }
 
-        public KnotEndpoint(string uri)
+        public KnotEndpoint(string uri, int timeout = 10)
         {
             _uri = uri;
+            _timeout = 10;
         }
     }
 }