+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.1.2 (2023-03-24)
+* Expose KnotRequest.WebRequest
+* Add endpoint request timeout parameter
+
 ## 0.1.0 (2023-03-23)
 * Initial release